move CVXOPT-only utility into cvxopt_conif.py
move tests of CVXOPT utility from free-function in test_conic_solvers.py to the class test_conic_solvers.py::TestCVXOPT
