+++ cvxpy/reductions/solvers/conic_solvers/cvxopt_conif.py
@@ -23,11 +23,103 @@ from scipy.sparse.linalg import eigsh
 import cvxpy.interface as intf
 import cvxpy.settings as s
 from cvxpy.constraints import PSD, SOC, NonNeg, Zero
-from cvxpy.reductions.solvers.compr_matrix import compress_matrix
 from cvxpy.reductions.solvers.conic_solvers.conic_solver import ConicSolver
 from cvxpy.reductions.solvers.kktsolver import setup_ldl_factor
 from cvxpy.reductions.solvers.openmp_conflict import warn_if_omp_conflict
 from cvxpy.utilities.citations import CITATION_DICT
+
+
+
+def compress_matrix(A, b, equil_eps: float = 1e-10):
+    """Compresses A and b by eliminating redundant rows.
+
+    Identifies rows that are positive multiples of another row,
+    i.e. rows that are redundant when (A, b) encodes Ax <= b.
+    Reduces A and b to C = PA, d = Pb, where P has one
+    nonzero per row.
+
+    Parameters
+    ----------
+    A : SciPy CSR matrix
+        The constraints matrix to compress.
+    b : NumPy 1D array
+        The vector associated with the constraints matrix.
+    equil_eps : float, optional
+        Standard for considering two numbers equivalent.
+
+    Returns
+    -------
+    tuple
+        The tuple (A, b, P, row_to_keep) where A and b are compressed
+        according to P and row_to_keep lists the retained row indices.
+    """
+    # Data for compression matrix.
+    P_V = []
+    P_I = []
+    P_J = []
+    # List of rows to keep.
+    row_to_keep = []
+    row_to_col = {}
+    # A map of sparsity pattern to row list.
+    sparsity_to_row = {}
+    prev_ptr = A.indptr[0]
+
+    def get_row_nnz(row):
+        return A.indptr[row+1] - A.indptr[row]
+
+    for row_num in range(A.shape[0]):
+        keep_row = True
+        ptr = A.indptr[row_num+1]
+        pattern = tuple(A.indices[prev_ptr:ptr])
+        # Eliminate empty rows.
+        nnz = ptr - prev_ptr
+        if nnz == 0 or np.linalg.norm(A.data[prev_ptr:ptr]) < equil_eps:
+            keep_row = False
+            P_V.append(0.)
+            P_I.append(row_num)
+            P_J.append(0)
+        # Sparsity pattern is the same or there was a false collision.
+        # Check rows have the same number of nonzeros.
+        elif pattern in sparsity_to_row and nnz == \
+                get_row_nnz(sparsity_to_row[pattern][0]):
+            # Now test if one row is a multiple of another.
+            row_matches = sparsity_to_row[pattern]
+            for row_match in row_matches:
+                cur_vals = A.data[prev_ptr:ptr]
+                prev_match_ptr = A.indptr[row_match]
+                match_ptr = A.indptr[row_match+1]
+                match_vals = A.data[prev_match_ptr:match_ptr]
+                # Ratio should be constant and positive. A negative
+                # multiple of an inequality row describes the opposite
+                # half-space, so it is not redundant.
+                ratio = cur_vals/match_vals
+                if ratio[0] > 0 and np.ptp(ratio) < equil_eps and \
+                   abs(ratio[0] - b[row_num]/b[row_match]) < equil_eps:
+                    keep_row = False
+                    P_V.append(ratio[0])
+                    P_I.append(row_num)
+                    P_J.append(row_to_col[row_match])
+            if keep_row:
+                sparsity_to_row[pattern].append(row_num)
+        # Pattern doesn't match anything present.
+        else:
+            sparsity_to_row[pattern] = [row_num]
+
+        if keep_row:
+            row_to_keep.append(row_num)
+            row_to_col[row_num] = len(row_to_keep)-1
+            P_V.append(1.)
+            P_I.append(row_num)
+            P_J.append(len(row_to_keep)-1)
+        prev_ptr = ptr
+
+    # Compress A and b.
+    cols = max(len(row_to_keep), 1)
+    P = sp.coo_array((P_V, (P_I, P_J)), shape=(A.shape[0], cols))
+    A_compr = A[row_to_keep, :]
+    b_compr = b[row_to_keep]
+    return (A_compr, b_compr, P, row_to_keep)
+
 
 
 # Utility method for formatting a ConeDims instance into a dictionary
